+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/chain/StatefulLink.kt
@@ -4,6 +4,6 @@ abstract class StatefulLink<K, V> : Link<V> {
     abstract val state: LinkStateManager<K, V>
     override val retryAfterBreak = true
     override val breakCondition: BreakCondition? = null
-    override val beforeAction: Action? = null
-    override val afterAction: Action? = null
+    override val beforeAction: ChainAction? = null
+    override val afterAction: ChainAction? = null
 }